requirements.txt
streamlit>=1.20
requests>=2.28
python-dotenv>=1.0
streamlit-autorefresh>=0.0.6
yfinance>=0.2.11
pandas>=1.5

